Use PNG files from the current shell directory
Instead of using PNG files relative to the JSON file passed in as an
argument, use PNG files that are in the current shell directory.

This is useful if we want to cd to different directories and apply the
same JSON to them. Without this change, we would need to have a multiple
copies of the same JSON file.